move simulcast utils to unifiedPlanUtils file
